Fix compilation warnings with -Wconversion

+++ esp32/main/byteops.h
@@ -69,7 +69,7 @@ static inline uint32_t extractle32(const uint8_t *source)
 static inline uint16_t extractbe16(const uint8_t *source)
 {
     uint16_t ret = ((((uint16_t)source[0]) << 8) +
-        ((uint16_t)source[1]));
+                    ((uint16_t)source[1]));
     return ret;
 }
 
@@ -83,8 +83,8 @@ static inline uint16_t extractbe16(const uint8_t *source)
  */
 static inline uint8_t *putle16(uint8_t *dest, uint16_t val)
 {
-    *dest++ = val & 0xff;
-    *dest++ = val >> 8;
+    *dest++ = (uint8_t)(val & 0xff);
+    *dest++ = (uint8_t)(val >> 8);
     return dest;
 }
 
@@ -97,8 +97,8 @@ static inline uint8_t *putle16(uint8_t *dest, uint16_t val)
  */
 static inline int putle16_c(uint8_t *dest, uint16_t val)
 {
-    *dest++ = val & 0xff;
-    *dest++ = val >> 8;
+    *dest++ = (uint8_t)(val & 0xff);
+    *dest++ = (uint8_t)(val >> 8);
     return 2;
 }
